FUNCTIONAL Implementation of the New Equation of State
The equation of state implemented in Si3D follows the definitions provided by
Gill (1982). The density of water is computed as a function of salinity (S), temperature (T), and
pressure (p). This is very important for Lake Tahoe (~500 m deep)
Several needs needed some changes in order to compile